+++ src/main/java/com/mraof/minestuck/blockentity/machine/SendificatorBlockEntity.java
@@ -1,5 +1,8 @@
 package com.mraof.minestuck.blockentity.machine;
 
+import com.mraof.minestuck.api.uranium.IUraniumHandler;
+import com.mraof.minestuck.api.uranium.SimpleUraniumHandlers;
+import com.mraof.minestuck.api.uranium.UraniumPower;
 import com.mraof.minestuck.blockentity.MSBlockEntityTypes;
 import com.mraof.minestuck.inventory.OptionalPosHolder;
 import com.mraof.minestuck.inventory.SendificatorMenu;
@@ -26,13 +29,13 @@ import net.neoforged.neoforge.items.wrapper.RangedWrapper;
 import javax.annotation.Nullable;
 import java.util.Optional;
 
-public class SendificatorBlockEntity extends MachineProcessBlockEntity implements MenuProvider, UraniumPowered
+public class SendificatorBlockEntity extends MachineProcessBlockEntity implements MenuProvider
 {
 	public static final String TITLE = "container.minestuck.sendificator";
 	public static final short MAX_FUEL = 128;
 	
 	private final ProgressTracker progressTracker = new ProgressTracker(ProgressTracker.RunType.ONCE_OR_LOOPING, 0, this::setChanged, this::contentsValid);
-	private short fuel = 0;
+	private final SimpleUraniumHandlers fuel = new SimpleUraniumHandlers.Insert(MAX_FUEL);
 	
 	@Nullable
 	private BlockPos destBlockPos;
@@ -42,13 +45,13 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 		@Override
 		public int get()
 		{
-			return fuel;
+			return fuel.getUraniumStored();
 		}
 		
 		@Override
 		public void set(int value)
 		{
-			fuel = (short) value;
+			fuel.setUraniumStored(value);
 		}
 	};
 	private final OptionalPosHolder destinationHolder = OptionalPosHolder.forPos(() -> Optional.ofNullable(this.getDestinationBlockPos()));
@@ -90,7 +93,14 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 			this.destBlockPos = new BlockPos(destX, destY, destZ);
 		}
 		
-		fuel = compound.getShort("fuel");
+		if(compound.contains("fuel", 99))
+		{
+			// Kept for backwards compatibility
+			fuel.setUraniumStored(compound.getShort("fuel"));
+		} else
+		{
+			fuel.load(compound.getCompound("fuel"));
+		}
 	}
 	
 	@Override
@@ -107,7 +117,7 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 			compound.putInt("destZ", destBlockPos.getZ());
 		}
 		
-		compound.putShort("fuel", fuel);
+		compound.put("fuel", fuel.save());
 	}
 	
 	@Override
@@ -131,7 +141,7 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 		
 		ItemStack fuel = itemHandler.getStackInSlot(1);
 		ItemStack input = itemHandler.getStackInSlot(0);
-		return canBeRefueled() && fuel.is(ExtraModTags.Items.URANIUM_CHUNKS) || !input.isEmpty();
+		return canBeRefueled(fuel) || !input.isEmpty();
 	}
 	
 	/**
@@ -139,13 +149,14 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 	 */
 	private void processContents()
 	{
-		if(canBeRefueled())
+		ItemStack fuel = itemHandler.getStackInSlot(1);
+		if(canBeRefueled(fuel))
 		{
 			//checks for a uranium itemstack in the lower(fuel) item slot, increases the fuel value if some is found and then removes one count from the fuel stack
-			if(itemHandler.getStackInSlot(1).is(ExtraModTags.Items.URANIUM_CHUNKS))
+			if(fuel.is(ExtraModTags.Items.URANIUM_CHUNKS))
 			{
 				//Refill fuel
-				addFuel((short) FUEL_INCREASE);
+				addFuel(fuel);
 				itemHandler.extractItem(1, 1, false);
 			}
 		}
@@ -166,7 +177,7 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 						ItemEntity itemEntity = new ItemEntity(level, destinationPos.getX(), destinationPos.getY(), destinationPos.getZ(), sentStack);
 						level.addFreshEntity(itemEntity);
 						
-						fuel = (short) (fuel - 8);
+						this.fuel.extractUranium(8, false);
 					}
 				}
 			}
@@ -178,27 +189,22 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 	 */
 	private boolean canSend()
 	{
-		return fuel > 0 && !itemHandler.getStackInSlot(0).isEmpty();
+		return fuel.getUraniumStored() > 0 && !itemHandler.getStackInSlot(0).isEmpty();
 	}
 	
 	/**
 	 * Checks that fuel can be added without any excess/wasted points being attributed
 	 */
-	public boolean canBeRefueled()
+	public boolean canBeRefueled(ItemStack fuelStack)
 	{
-		return fuel <= MAX_FUEL - FUEL_INCREASE;
+		int amount = UraniumPower.getUraniumPower(fuelStack);
+		return fuel.receiveUranium(amount, true) == amount;
 	}
 	
-	@Override
-	public boolean atMaxFuel()
+	public void addFuel(ItemStack fuelStack)
 	{
-		return fuel >= MAX_FUEL;
-	}
-	
-	@Override
-	public void addFuel(short fuelAmount)
-	{
-		fuel += fuelAmount;
+		int amount = UraniumPower.getUraniumPower(fuelStack);
+		fuel.receiveUranium(amount, false);
 	}
 	
 	@Nullable
@@ -212,6 +218,12 @@ public class SendificatorBlockEntity extends MachineProcessBlockEntity implement
 		if(side == Direction.DOWN)
 			return null;
 		return new RangedWrapper(this.itemHandler, 1, 2);
+	}
+	
+	@Nullable
+	public IUraniumHandler getUraniumHandler(@Nullable Direction side)
+	{
+		return fuel;
 	}
 	
 	public void openMenu(ServerPlayer player)